requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
edge-tts>=6.1.9
pygame>=2.5.2
google-genai>=0.2.0
requests>=2.31.0
vosk>=0.3.32
pyaudio>=0.2.13
pvrecorder>=1.2.1
groq>=0.4.0
deepgram-sdk>=3.0.0
python-dotenv>=1.0.0
